Enhanced VC-1 Acceleration Options Provide enhanced user configurable options for VC-1 hardware acceleration, based on selection criteria
Currently there is no option to configure hardware acceleration for VC-1 playback. This presents issues in some instances such as hardware accelerated VC-1 decoding is subject to forced de-interlacing performed on even progressive content, resulting in considerable resolution loss for this content. It appears hardware such as the S922X SoCs are powerful enough to perform software decoding of even 1080p24 content using the ffmpeg libraries for VC-1 playback.

Options are: Never, Always, HD and up, and Exclude 24p. Defaults to Always (to retain existing behavior in lieu of user setting this option).